Add Search class with tests
