+++ testdata/imagetest/info.json
@@ -1,0 +1,38 @@
+{
+    "@context": "http://iiif.io/api/image/2/context.json",
+    "@id": "imagetest",
+    "height": 3282,
+    "profile": [
+        "http://iiif.io/api/image/2/level0.json",
+        {
+            "supports": [
+                "cors",
+                "sizeByWhListed",
+                "baseUriRedirect"
+            ]
+        }
+    ],
+    "protocol": "http://iiif.io/api/image",
+    "sizes": [
+        {
+            "height": 3282,
+            "width": 1784
+        },
+        {
+            "height": 250,
+            "width": 136
+        }
+    ],
+    "tiles": [
+        {
+            "scaleFactors": [
+                1,
+                2,
+                4,
+                8
+            ],
+            "width": 512
+        }
+    ],
+    "width": 1784
+}
